fixed number of days requested when applying for a leave to be accurate

+++ src/main/java/com/jp/lms/dto/response/LeaveResponse.java
@@ -14,7 +14,7 @@ import java.time.LocalDate;
 public record LeaveResponse(
         LocalDate startDate,
         LocalDate endDate,
-        Integer numOfDaysRequested,
+        Long numOfDaysRequested,
         String handoverTo,
         String reason,
         LeaveType leaveType,
